fix CostCenterBreakupResponseDto Long to BigDecimal

+++ src/main/java/com/asg/finance/service/impl/BankDebitVoucherServiceImpl.java
@@ -1044,7 +1044,7 @@ public class BankDebitVoucherServiceImpl implements BankDebitVoucherService {
                                     dto.setCostPoid(cc.getCostPoid());
                                     dto.setAmount(
                                             cc.getAmount() != null
-                                                    ? BigDecimal.valueOf(cc.getAmount())
+                                                    ? (cc.getAmount())
                                                     : BigDecimal.ZERO
                                     );
                                     if (cc.getCostPoid() != null && !cc.getCostPoid().isEmpty() &&
